KeyUI and GameObject
added the key UI; hit Q to turn the image true or false.
added the key gameobject and gave it a box collider. attached a script to destory the game object if collides with anything. It is on the same sorting layer as the player and the same z postion but the collision is not happening/